pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.70.0",
    "python-dotenv==1.1.0",
]
